Add overlap detection
The program will detect if the edge will overlap, it will not route if there is overlap. Hence, test4 and test5 cannot be route as there are overlap. It did not check the capacity. The capacity checking and overlap allowing within the capacity will be done later so that test4 and test5 can be routed.